fix:前端联调-直播间IM服务

+++ yubao-live-living-provider/src/main/java/cn/junbao/yubao/live/living/provider/dao/mapper/ILivingRoomMapper.java
@@ -11,13 +11,13 @@ public interface ILivingRoomMapper {
 
     void insertOne(LivingRoomPO livingRoomPO);
 
-    LivingRoomPO selectByRoomId(@Param("roomId") Long roomId);
+    LivingRoomPO selectByRoomId(@Param("roomId") Integer roomId);
 
-    void deleteByRoomId(Long roomId);
+    void deleteByRoomId(Integer roomId);
 
     List<LivingRoomPO> selectLivingRoomPage(@Param("type") Integer type,@Param("offset") int offset, @Param("limit") int limit);
 
     List<LivingRoomPO>  selectAll(@Param("type") Integer type);
 
-    LivingRoomPO selectOneByRoomId(@Param("roomId") Long roomId);
+    LivingRoomPO selectOneByRoomId(@Param("roomId") Integer roomId);
 }
